### 2020-02-21 #2110 @whedon
[PRE REVIEW]: Autorank: A Python package for automated ranking of classifiers
**Submitting author:** @sherbold (<a href="http://orcid.org/0000-0001-9765-2803">Steffen Herbold</a>)
**Repository:** <a href="https://github.com/sherbold/autorank" target ="_blank">https://github.com/sherbold/autorank</a>
**Version:** 1.0.0
