### 2020-02-21 #2110 @lorenanicole
Greetings @kyleniemeyer ! I am traveling tomorrow to arrive early to be onsite for work next week! In terms of time expectations, I presume starting to work on this next weekish is agreeable?
