### 2020-04-09 #2110 @arfon
> Thanks for the info. I would be happy with any progress, but understand waiting as well. My own workload is also quite heavily impacted already.

@sherbold - I'm going to take over the handling of this submission for now.
When you have a chance, could you take a look a [this list of potential reviewers](https://bit.ly/joss-reviewers) and identify a few people who would be good candidates to review this submission?

Thanks!
