### 2020-04-10 #2110 @sherbold
Thanks for taking over. I just went from the top and looked for reviewers that have statistics + something related to ML or data analysis as domain, Python as preferred language and no active reviews. Here are eight candidates, but I only went half-way through the list and could look for if they do not have time.

- betatim
- tupi
- tianxzhu
- dirmeier
- adavidzh
- JonathanReardon
- ejolly
- bahung
- mskwark
