### 2020-04-10 #2110 @arfon
@JonathanReardon @bahung @ejolly -
would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The package we would ask you to review is [_Autorank: A Python package for automated ranking of classifiers_](https://github.com/sherbold/autorank) and there is a [short associated paper](https://github.com/openjournals/joss-papers/blob/joss.02110/joss.02110/10.21105.joss.02110.pdf) to read too.
