### 2020-04-11 #2173 @ejhigson
Hi @sherbold - this is a really nice piece of software. I really like the idea and definitely see the practical usefulness!

Installing your package and starting playing around with it, I noticed a few things which could be changed or fixed - I will comment about them below.
