### 2020-04-11 #2173 @ejhigson
1) I think you need to add version requirements for you dependencies as I got errors running the example at https://sherbold.github.io/autorank/ which were fixed by upgrading the dependency module versions. E.g. with scipy 1.1 I get the following error:
```
AttributeError: module 'scipy.stats' has no attribute 'median_absolute_deviation'
```
Upgrading scipy to 1.4.1 fixed the issue. Perhaps you can consider specifying minimum version requirements for your dependencies in setup.py? E.g. "scipy>=1.4" or whatever the appropriate version is.
