### 2020-04-11 #2173 @ejhigson
2. Following this step in the example usage at https://sherbold.github.io/autorank/#documentation:
```
plot_stats(res)
plt.show()
```
I get the following error:
```
RuntimeError: latex was not able to process the following string:
b'pop_0'
```
I see similar errors when I try running the module tests with ``nosetests``. Do you know what is causing these errors?

You may also want to change the code in this cell at https://sherbold.github.io/autorank/#documentation to use the variable name "results" instead of "res" for consistency with earlier cells, and to have an "import matplotlib.pyplot as plt" statement.

For reference, I am using python 3.6.9 with these module versions:
```
matplotlib                        3.2.1
pandas                            1.0.3
scipy                             1.4.1
```

My laptop is running Ubuntu 18, and my latex version is:
```
$ latex --version
pdfTeX 3.14159265-2.6-1.40.18 (TeX Live 2017/Debian)
kpathsea version 6.2.3
Copyright 2017 Han The Thanh (pdfTeX) et al.
There is NO warranty.  Redistribution of this software is
covered by the terms of both the pdfTeX copyright and
the Lesser GNU General Public License.
For more information about these matters, see the file
named COPYING and the pdfTeX source.
Primary author of pdfTeX: Han The Thanh (pdfTeX) et al.
Compiled with libpng 1.6.34; using libpng 1.6.34
Compiled with zlib 1.2.11; using zlib 1.2.11
Compiled with poppler version 0.62.0
```
