### 2020-04-11 #2173 @ejhigson
3. I think there may be a typo in your flow chart as both arrows leaving the "Two populations?" diamond say "Yes". Should one of them say "No"?
