### 2020-04-11 #2173 @ejhigson
3. Regarding the paper itself: I think there is a typo in the last sentence of the third paragraph "Instead, additional aspect such as effect sizes, confidence intervals, and the decision whether it is appropriate to report the  mean value and standard deviation, or whether the median value and the median absolute deviation are better suited". Also you may wish to add DOIs to your references.
