### 2020-04-11 0003513
Specify required versions for dependencies.
No version for numpy specified, because only very basic functions are needed and the other dependencies likely have stricter requirements on numpy anyways.

Closes #1
