### 2020-04-11 3d061c4
Fixed wrong usage of res instead of result in the documentation
