### 2020-04-11 f6bd818
Add matplotlib import to documentation
