### 2020-04-11 #2173 @sherbold
@ejhigson Thanks for the good feedback. I fixed 1, 3, and 4. Unfortunately, neither arXiv, nor JLMR have DOIs. Why this is a case for JLMR is a mystery for me.

Regarding 2: I fixed the documentation. They are not updated yet, because GH Actions seem to be broken right now, though there is no incident report filed yet.

However, I cannot replicate any test failures. I ran the tests under Windows and Ubuntu 18.04 and did not see anything. I am waiting for the GH actions to become live again, there tests are running there as well.

Or do you see the error, when you try to compile the generated latex code?
