### 2020-04-11 3be4f51
Remove unused imports in sklearn example
