### 2020-04-11 eb1ef72
Remove support for Python 3.5
