+++ .github/workflows/publish-to-test-pypi.yml
@@ -8,7 +8,7 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-18.04, windows-latest]
-        python: ['3.5.x', '3.6.x', '3.7.x', '3.8.x']
+        python: ['3.6.x', '3.7.x', '3.8.x']
     runs-on: ${{ matrix.os }}
     steps:
     - uses: actions/checkout@master
