### 2020-04-11 c6bf2e3
Upgrade to latest pydocmd version
Unfortunately, there were breaking changes from 2.0.5 tp 2.1.0. Now
we use the latest version and fixed the version number in the build.
