+++ .github/workflows/publish-to-test-pypi.yml
@@ -31,7 +31,7 @@ jobs:
         python setup.py sdist bdist_wheel
     - name: Build documentation
       run: |
-        python -m pip install pydoc-markdown
+        python -m pip install pydoc-markdown==2.1.3
         mkdir docs
         mkdir docs/examples
         cp examples/* docs/examples/
