### 2020-04-11 #2173 @sherbold
The documentation is now also updated. I also did some maintenance of the Github Action and documentation build, because there were breaking changes in the latest releases.
