### 2020-04-12 01a246f
Added a newline to fix markdown formatting in documentation
Github shows this correctly without the newline.
