### 2020-04-12 #2173 @ejhigson
> However, I cannot replicate any test failures. I ran the tests under Windows and Ubuntu 18.04 and did not see anything. I am waiting for the GH actions to become live again, there tests are running there as well.
>
> Or do you see the error, when you try to compile the generated latex code?

@sherbold I found the source of the problem. The error was in the creation of matplotlib figures not the generated latex. I had specified the matplotlib parameter text.usetex as True in my matplotlibrc but it seems your example and tests require text.usetex = False to run. With this matplotlib setting then everything works perfectly for me. You could perhaps consider adding a statement to check the matplotlib rc parameter text.usetex is set to False and, if not, to issue a UserWarning or similar to tell people they need to change it.
