### 2020-04-12 #2173 @ejhigson
Here are a few other minor suggestions before I finish my review (sorry any of the below already exist and I have missed them):
1) please could you document how to run your tests in your documentation page? I tried running with nosetests and everything worked great but I think it is worth stating how to do this.
2) the API documentation in the docstrings is very good. Perhaps its worth explicitly suggesting in your documentation page that people look at it for more details?
3) One of the review criteria for the paper itself is "State of the field: Do the authors describe how this software compares to other commonly-used packages?" I don't think you mention any other similar software packages. Do any exist? If not then this is fine.
