### 2020-04-12 e07dfac
Added dev guide to documentation
Also changed documentation link to directly point to the API doc.
