### 2020-04-12 5be11f0
Fix a couple of typos.
