### 2020-04-12 558f990
Added UserWarning in case text.usetex is True
