### 2020-04-12 45679db
Added sentence to differentiate autorank from other packages by a comparison with scipy.
