### 2020-04-12 #2173 @sherbold
Thanks for the additional comments. I addressed them as follows:

- I added a UserWarning that shows when autorank is important if text.usetex==True. I am thinking about adding an additional UserWarning to plot_stats as well. But I will likely find the source for the problem with usetex at some point.
- I added a short guide for devs, including test execution: https://github.com/sherbold/autorank#creating-a-local-developer-environment
- I changed the link to specifically point to the API-Doc. The remainder of the documentation is the README anyways.
- I added a statement how autorank is different from other statistical packages using scipy as example.
