### 2020-04-14 #2173 @JonathanReardon
There is also a missing word in the 'output' example (https://sherbold.github.io/autorank/). In the last sentence I think it should be 'Based on the post-hoc Nemenyi test' rather than 'Based the post..'.
