### 2020-04-14 90947e8
Beautified string output of RankResult and _ComparisonResult
