### 2020-04-14 b5c2415
Fixed typo in output and example
