### 2020-04-14 #2173 @sherbold
Thank you @JonathanReardon . I have overwritten the namedtuple with a customized __str__ function. The output now uses newlines and looks as follows:
```
RankResult(rankdf=
       meanrank    median       mad  ci_lower  ci_upper  effect_size  \
pop_5      2.18  0.912005  0.130461  0.692127         1  2.66454e-17
pop_4      2.29  0.910437  0.132786  0.654001         1       -0.024
pop_3      2.47  0.858091  0.210394  0.573879         1       0.1364
pop_2      3.95  0.505057  0.333594  0.227184   0.72558       0.6424
pop_1      4.71  0.313824  0.247339  0.149473  0.546571       0.8516
pop_0      5.40  0.129756  0.192377         0  0.349014       0.9192

        magnitude
pop_5  negligible
pop_4  negligible
pop_3  negligible
pop_2       large
pop_1       large
pop_0       large
pvalue=2.3412212612346733e-28
cd=1.0662484349869374
omnibus=friedman
posthoc=nemenyi
all_normal=False
pvals_shapiro=[1.646607051952742e-05, 0.0605173334479332, 0.13884511590003967, 0.00010030837438534945, 2.066387423838023e-06, 1.5319776593969436e-06]
homoscedastic=True
pval_homogeneity=0.2663177301695518
homogeneity_test=levene
alpha=0.05
alpha_normality=0.008333333333333333
num_samples=50)
```
I also fixed the typo with the Nemenyi test.

I did not find the missing 'output'. Could you tell me exactly where to look? I seem to be typo-blind.
