### 2020-04-14 #2173 @JonathanReardon
Thank you for making these changes @sherbold. The missing word in the example 'create_report(result)' output was the part you already corrected, sorry for the slight lack of clarity there on my part :)

My review is now complete, all boxes ticked. Congratulations on a great piece of software, I look forward to seeing it published and widely used.

Cheers
