### 2020-04-14 #2173 @sherbold
Thank you both @ejhigson and @JonathanReardon for the fast reviews and the helpful comments that really improved the tool and documentation again. The changes were small, but the impact in terms of user experience should be large :)
