+++ paper/paper.bib
@@ -43,5 +43,5 @@
       volume={17},
       pages={261--272},
       adsurl = {https://rdcu.be/b08Wh},
-      doi = {https://doi.org/10.1038/s41592-019-0686-2},
+      doi = {10.1038/s41592-019-0686-2},
 }
