### 2020-04-14 #2173 @arfon
@sherbold - could you merge this fix to one of the DOIs? https://github.com/sherbold/autorank/pull/2
