### 2020-04-14 #2173 @sherbold
Version 1.0.1 with all changes is released and available on PyPi: https://pypi.org/project/autorank/1.0.1/

There is also a long-term archive of 1.0.1 on Zenodo:
https://doi.org/10.5281/zenodo.3751301

@arfon Thank you for taking over and the fast process. I also want to stress how much I like the approach by JOSS and look forward to contributing back to the journal at some point, when things get back to (more) normal :)
