### 2020-04-14 #2173 @arfon
> @arfon Thank you for taking over and the fast process. I also want to stress how much I like the approach by JOSS and look forward to contributing back to the journal at some point, when things get back to (more) normal :)

Sure thing!
